Add Types.h and SafeCast

+++ AnimBoxCore/src/AnimBoxCore/Memory/Private/ArenaAllocator.cpp
@@ -1,27 +1,27 @@
 ﻿#include "PCH.h"
 #include "AnimBoxCore/Memory/Public/ArenaAllocator.h"
 
-AnimBox::ArenaAllocator::ArenaAllocator(size_t capacity)
+AnimBox::ArenaAllocator::ArenaAllocator(uint64 capacity)
 {
     _buffer = static_cast<uint8_t*>(malloc(capacity));
     _capacity = capacity;
     _offset = 0;
 }
 
-void* AnimBox::ArenaAllocator::Allocate(size_t size, size_t alignment)
+void* AnimBox::ArenaAllocator::Allocate(uint64 size, uint64 alignment)
 {
     const uint8_t* currentPtr = (_buffer + _offset);
 
-    const size_t currentAddress = reinterpret_cast<size_t>(currentPtr);
+    const uint64 currentAddress = reinterpret_cast<uint64>(currentPtr);
     // round up the pointer to next multiple of alignment
-    const size_t alignedAddress = (currentAddress + alignment - 1) & ~(alignment - 1);
+    const uint64 alignedAddress = (currentAddress + alignment - 1) & ~(alignment - 1);
 
-    const size_t padding = alignedAddress - currentAddress;
+    const uint64 padding = alignedAddress - currentAddress;
 
     ASSERT(_offset + size + padding <= _capacity, "Arena Allocator overflow")
 
     _offset += padding;
-    uint8_t* alignedPtr = _buffer + _offset;
+    uint8* alignedPtr = _buffer + _offset;
     
     _offset += size;
     return alignedPtr;
