+++ AnimBoxCore/src/AnimBoxCore/Memory/Public/Memory.h
@@ -1,16 +1,17 @@
 ﻿#pragma once
 
 #include <cstdint>
+#include "AnimBoxCore/Core/Public/Types.h"
 
 namespace AnimBox
 {
     struct AllocationMetrics
     {
-        uint32_t TotalAllocated = 0;
-        uint32_t TotalFreed = 0;
+        uint32 TotalAllocated = 0;
+        uint32 TotalFreed = 0;
 
         void LogAllocationMetrics() const;
-        uint32_t GetCurrentUsage() const { return TotalAllocated - TotalFreed; }
+        uint32 GetCurrentUsage() const { return TotalAllocated - TotalFreed; }
 
         static AllocationMetrics& GetHeapAllocationMetrics()
         {
@@ -20,5 +21,5 @@ namespace AnimBox
     };
 }
 
-void* operator new(size_t size);
-void operator delete(void* memory, size_t size);
+void* operator new(uint64 size);
+void operator delete(void* memory, uint64 size);
